requirements.txt
fastapi
uvicorn
python-dotenv
smolagents
litellm
pandas
numpy
matplotlib
geopandas
rasterio
geedim
tslearn
scikit-learn
requests
openpyxl
pypdf
earthengine-api
pydantic
langfuse
